working .apk with tests

+++ UC_Midterm/app/src/main/java/com/example/unitcoverter/Converter.java
@@ -1,10 +1,8 @@
 package com.example.unitcoverter;
 
 public class Converter {
-    public static float toCelsius(float fah) {
-        return (fah - 32) * 5/9;
-    }
-    public static float toKiloG(float lbs) {return (lbs * (float)0.45359237); }
-    public static float toML(float oz) {return (oz * (float)29.57); }
-    public static float toKM(float miles) {return (miles * (float)1.60934);}
+    public static double toCelsius(double fah) {return (fah - 32) * 5/9;}
+    public static double toKiloG(double lbs) {return (lbs * 0.45359237);}
+    public static double toML(double oz) {return (oz * 29.57);}
+    public static double toKM(double miles) {return (miles * 1.60934);}
 }
